Move Transport Node tests out of androidTest NetworkCardTest
Mirror the src/test changes on the androidTest side — CI's Instrumented
Tests job was failing because app/src/androidTest/.../NetworkCardTest.kt
still called NetworkCard with transportNodeEnabled / onTransportNodeToggle
after those params were removed.

Deletes the five transport-node tests from the androidTest NetworkCardTest
and moves them to a new androidTest AdvancedCardTest, targeting AdvancedCard.
Also fixes networkCard_withAllDefaultValues_displaysCorrectly to drop its
Transport Node assertion.
